finish with solution that works only on sparse tensor. Slower then CPU. Need to speed up multiplication

+++ TTCross/include/core.cuh
@@ -9,18 +9,18 @@
 class Core {
 private:
     size_t a_, b_, c_;
-    float* data_{};
-    std::vector<DevMatrix<float>> matrices;
+    double* data_{};
+    std::vector<DevMatrix<double>> matrices;
 
 public:
     Core();
-    Core(const DevMatrix<float>& A,  size_t r_prev, size_t n_k, size_t r_k);
+    Core(const DevMatrix<double>& A,  size_t r_prev, size_t n_k, size_t r_k);
     Core(const TMatrix& A, size_t r_prev, size_t n_k, size_t r_k);
     Core(const Core& other);
     ~Core();
     std::tuple<size_t,size_t,size_t> Sizes() const;
-    void SetMatrices(const DevMatrix<float>& A,  size_t r_prev, size_t n_k, size_t r_k);
-    DevMatrix<float> Matrix(size_t i) const;
+    void SetMatrices(const DevMatrix<double>& A,  size_t r_prev, size_t n_k, size_t r_k);
+    DevMatrix<double> Matrix(size_t i) const;
 
     double operator()(size_t i, size_t j, size_t k) const;
     Core& operator= (const Core& other);
