+++ TTCross/src/host_matrix.cu
@@ -5,16 +5,16 @@
 Matrix::Matrix() : rows_(0), cols_(0), data_(nullptr) {}
 
 
-Matrix::Matrix(size_t rows, size_t cols, float val) : rows_(rows), cols_(cols) {
-    size_t bytes = sizeof(float) * rows_ * cols_;
+Matrix::Matrix(size_t rows, size_t cols, double val) : rows_(rows), cols_(cols) {
+    size_t bytes = sizeof(double) * rows_ * cols_;
 
-    data_ = (float*)malloc(bytes);
+    data_ = (double*)malloc(bytes);
 }
 
 
 Matrix::Matrix(const Matrix& other) : rows_(other.rows_), cols_(other.cols_) {
-    size_t bytes = sizeof(float) * rows_ * cols_;
-    data_ = (float*)malloc(bytes);
+    size_t bytes = sizeof(double) * rows_ * cols_;
+    data_ = (double*)malloc(bytes);
     memcpy(data_, other.data_, bytes);
 }
 
@@ -37,9 +37,9 @@ Matrix& Matrix::operator= (const Matrix& other) {
     rows_ = other.rows_;
     cols_ = other.cols_;
 
-    size_t bytes = sizeof(float) * rows_ * cols_;
+    size_t bytes = sizeof(double) * rows_ * cols_;
 
-    data_ = (float*)malloc(bytes);
+    data_ = (double*)malloc(bytes);
 
     memcpy(data_, other.data_, bytes);
 
@@ -47,12 +47,12 @@ Matrix& Matrix::operator= (const Matrix& other) {
 }
 
 
-float& Matrix::operator() (size_t i, size_t j) {
+double& Matrix::operator() (size_t i, size_t j) {
     return data_[IDX2C(i,j,rows_)];
 }
 
 
-float Matrix::operator() (size_t i, size_t j) const {
+double Matrix::operator() (size_t i, size_t j) const {
     return data_[IDX2C(i,j,rows_)];
 }
 
@@ -78,11 +78,11 @@ size_t Matrix::Cols() const {
     return cols_;
 }
 
-float *Matrix::Data() {
+double *Matrix::Data() {
     return data_;
 }
 
-const float *Matrix::Data() const {
+const double *Matrix::Data() const {
     return data_;
 }
